# Optimize ARM decoder further