I've done some more work on the ARM decoder, unrolling the match copy loop to get a few % better throughput:

    name                old time/op    new time/op    delta
    Uncompress-4          39.0ns ± 1%    40.1ns ± 2%  +2.82%  (p=0.000 n=19+10)
    UncompressPg1661-4    3.75ms ± 1%    3.54ms ± 2%  -5.83%  (p=0.000 n=18+9)
    UncompressDigits-4     246µs ± 1%     237µs ± 1%  -3.51%  (p=0.000 n=20+9)
    UncompressTwain-4     2.31ms ± 3%    2.16ms ± 2%  -6.14%  (p=0.000 n=20+10)
    UncompressRand-4      14.2µs ± 4%    14.3µs ± 3%    ~     (p=0.130 n=18+10)
    
    name                old speed      new speed      delta
    UncompressPg1661-4   101MB/s ± 1%   107MB/s ± 2%  +6.20%  (p=0.000 n=18+9)
    UncompressDigits-4   387MB/s ± 1%   401MB/s ± 1%  +3.64%  (p=0.000 n=20+9)
    UncompressTwain-4    111MB/s ± 3%   119MB/s ± 2%  +6.56%  (p=0.000 n=20+10)
    UncompressRand-4    1.16GB/s ± 4%  1.15GB/s ± 3%    ~     (p=0.133 n=18+10)